Identify current Page + Output to CSV file
At this point in my website, I cannot figure out how to identify the current webpage with a id or class, everything I try in my .css is not working. Similarly, I have tried to use both python/javascript and javascript/php to write my form data to a csv but it is not going anywhere.